pa-workflow/backend/ocr/requirements.txt
paddlepaddle
paddleocr
flask
flask-cors
pdf2image
python-docx
Pillow
openpyxl
python-pptx